[tools] perf pmu: Support wildcards on pmu name in dynamic pmu events
Message-id: <20190211232742.16489-77-jolsa@redhat.com>
Patchwork-id: 241482
O-Subject: [PATCH RHEL7 BZ1672651 076/271] perf pmu: Support wildcards on pmu name in dynamic pmu events
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit b2b9d3a3f0211c5d08c7befdf9d4adad48cda315
Author: Agustin Vega-Frias <agustinv@codeaurora.org>
Date: Tue Mar 6 09:04:42 2018 -0500

description
===========
Starting on v4.12 event parsing code for dynamic pmu events already
supports prefix-based matching of multiple pmus when creating dynamic
events. E.g., in a system with the following dynamic pmus:

    mypmu_0
    mypmu_1
    mypmu_2
    mypmu_4

passing mypmu/<config>/ as an event spec will result in the creation of
the event in all of the pmus. This change expands this matching through
the use of fnmatch so glob-like expressions can be used to create events
in multiple pmus. E.g., in the system described above if a user only
wants to create the event in mypmu_0 and mypmu_1, mypmu_[01]/<config>/
can be passed.

Change-Id: Icb25653fc5d5239c20f3bffdfdf4ab4c9c9bb20b
Signed-off-by: Jan Stancek <jstancek@redhat.com>