[tools] perf intel-pt: Fix sync_switch
Message-id: <20190211232742.16489-81-jolsa@redhat.com>
Patchwork-id: 241469
O-Subject: [PATCH RHEL7 BZ1672651 080/271] perf intel-pt: Fix sync_switch
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 63d8e38f6ae6c36dd5b5ba0e8c112e8861532ea2
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Mar 7 16:02:22 2018 +0200

description
===========
sync_switch is a facility to synchronize decoding more closely with the
point in the kernel when the context actually switched.

The flag when sync_switch is enabled was global to the decoding, whereas
it is really specific to the CPU.

The trace data for different CPUs is put on different queues, so add
sync_switch to the intel_pt_queue structure and use that in preference
to the global setting in the intel_pt structure.

That fixes problems decoding one CPU's trace because sync_switch was
disabled on a different CPU's queue.

Signed-off-by: Jan Stancek <jstancek@redhat.com>