[tools] perf intel-pt: Fix error recovery from missing TIP packet
Message-id: <20190211232742.16489-82-jolsa@redhat.com>
Patchwork-id: 241472
O-Subject: [PATCH RHEL7 BZ1672651 081/271] perf intel-pt: Fix error recovery from missing TIP packet
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 1c196a6c771c47a2faa63d38d913e03284f73a16
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Mar 7 16:02:23 2018 +0200

description
===========
When a TIP packet is expected but there is a different packet, it is an
error. However the unexpected packet might be something important like a
TSC packet, so after the error, it is necessary to continue from there,
rather than the next packet. That is achieved by setting pkt_step to
zero.

Signed-off-by: Jan Stancek <jstancek@redhat.com>