[tools] perf annotate: Fix s390 target function disassembly
Message-id: <20190211232742.16489-89-jolsa@redhat.com>
Patchwork-id: 241483
O-Subject: [PATCH RHEL7 BZ1672651 088/271] perf annotate: Fix s390 target function disassembly
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 0b58a77ca8792bd7798ef3a4d865c57694ec74fc
Author: Thomas Richter <tmricht@linux.vnet.ibm.com>
Date: Wed Mar 7 14:43:25 2018 +0100

description
===========
'perf annotate' displays function call assembler instructions with a
right arrow. Hitting enter on this line/instruction causes the browser
to disassemble this target function and show it on the screen.  On s390
this results in an error message 'The called function was not found.'

The function call assembly line parsing does not handle the s390 bras
and brasl instructions. Function call__parse expects the target as first
operand:

	callq	e9140 <__fxstat>

S390 has a register number as first operand:

	brasl	%r14,41d60 <abort>

Therefore the target addresses on s390 are always zero which is an
invalid address.

Introduce a s390 specific call parsing function which skips the first
operand on s390.

Signed-off-by: Jan Stancek <jstancek@redhat.com>