[tools] perf vendor events: Fix error code in json_events()
Message-id: <20190211232742.16489-116-jolsa@redhat.com>
Patchwork-id: 241530
O-Subject: [PATCH RHEL7 BZ1672651 115/271] perf vendor events: Fix error code in json_events()
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 931ef5dc5c18717d24e5b8d8a968e35638508051
Author: John Garry <john.garry@huawei.com>
Date: Thu Mar 8 18:58:27 2018 +0800

description
===========
When EXPECT macro fails an assertion, the error code is not properly set
after the first loop of tokens in function json_events().

This is because err is set to the return value from func function
pointer call, which must be 0 to continue to loop, yet it is not reset
for for each loop. I assume that this was not the intention, so change
the code so err is set appropriately in EXPECT macro itself.

In addition to this, the indention in EXPECT macro is tidied. The
current indention alludes that the 2 statements following the if
statement are in the body, which is not true.

Signed-off-by: Jan Stancek <jstancek@redhat.com>