[tools] perf vendor events: Drop support for unused topic directories
Message-id: <20190211232742.16489-117-jolsa@redhat.com>
Patchwork-id: 241502
O-Subject: [PATCH RHEL7 BZ1672651 116/271] perf vendor events: Drop support for unused topic directories
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 6f2f2ca3454ec4fa03fcd4507bdd7fe97303065b
Author: John Garry <john.garry@huawei.com>
Date: Thu Mar 8 18:58:28 2018 +0800

description
===========
Currently a topic subdirectory is supported in the pmu-events dir, in
the following sample structure: /arch/platform/subtopic/mysubtopic.json

Upto 256 levels of topic subdirectories are supported. So this means
that JSONs may be located in a topic dir as well as the platform dir.

This topic subdirectory causes problems if we want to add support for a
vendor dir in the pmu-events structure (in the form
arch/platform/vendor), in that we cannot differentiate between a vendor
dir and a topic dir.

Since the topic dir feature is not used, drop it so it does not block
adding vendor subdirectory support.

Signed-off-by: Jan Stancek <jstancek@redhat.com>