[tools] perf tools: Fix python extension build for gcc 8
Message-id: <20190211232742.16489-141-jolsa@redhat.com>
Patchwork-id: 241525
O-Subject: [PATCH RHEL7 BZ1672651 140/271] perf tools: Fix python extension build for gcc 8
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit b7a313d84e853049062011d78cb04b6decd12f5c
Author: Jiri Olsa <jolsa@kernel.org>
Date: Mon Mar 19 09:29:02 2018 +0100

description
===========
The gcc 8 compiler won't compile the python extension code with the
following errors (one example):

  python.c:830:15: error: cast between incompatible  function types from              \
  PyObject * (*)(struct pyrf_evsel *, PyObject *, PyObject *)                       \
  uct _object * (*)(struct pyrf_evsel *, struct _object *, struct _object *)} to     \
  PyObject * (*)(PyObject *, PyObject *) {aka struct _object * (*)(struct _objeuct \
  _object *)} [-Werror=cast-function-type]
     .ml_meth  = (PyCFunction)pyrf_evsel__open,

The problem with the PyMethodDef::ml_meth callback is that its type is
determined based on the PyMethodDef::ml_flags value, which we set as
METH_VARARGS | METH_KEYWORDS.

That indicates that the callback is expecting an extra PyObject* arg, and is
actually PyCFunctionWithKeywords type, but the base PyMethodDef::ml_meth type
stays PyCFunction.

Previous gccs did not find this, gcc8 now does. Fixing this by silencing this
warning for python.c build.

Commiter notes:

Do not do that for CC=clang, as it breaks the build in some clang
versions, like the ones in fedora up to fedora27:

  fedora:25:error: unknown warning option '-Wno-cast-function-type'; did you mean '-Wno-bad-function-cast'? [-Werror,-Wunknown-warning-option]
  fedora:26:error: unknown warning option '-Wno-cast-function-type'; did you mean '-Wno-bad-function-cast'? [-Werror,-Wunknown-warning-option]
  fedora:27:error: unknown warning option '-Wno-cast-function-type'; did you mean '-Wno-bad-function-cast'? [-Werror,-Wunknown-warning-option]
  #

those have:

  clang version 3.9.1 (tags/RELEASE_391/final)

The one in rawhide accepts that:

  clang version 6.0.0 (tags/RELEASE_600/final)

Signed-off-by: Jan Stancek <jstancek@redhat.com>