[tools] perf probe: Use right type to access array elements
Message-id: <20190211232742.16489-144-jolsa@redhat.com>
Patchwork-id: 241526
O-Subject: [PATCH RHEL7 BZ1672651 143/271] perf probe: Use right type to access array elements
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit d0461794a1dcaf552b507e23788777f718b736a1
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Sat Mar 17 21:52:25 2018 +0900

description
===========
Current 'perf probe' converts the type of array-elements incorrectly. It
always converts the types as a pointer of array. This passes the "array"
type DIE to the type converter so that it can get correct "element of
array" type DIE from it.

E.g.
  ====
  $ cat hello.c
  #include <stdio.h>

  void foo(int a[])
  {
	  printf("%d\n", a[1]);
  }

  void main()
  {
	  int a[3] = {4, 5, 6};
	  printf("%d\n", a[0]);
	  foo(a);
  }

  $ gcc -g hello.c -o hello
  $ perf probe -x ./hello -D "foo a[1]"
  ====

Without this fix, above outputs
  ====
  p:probe_hello/foo /tmp/hello:0x4d3 a=+4(-8(%bp)):u64
  ====
The "u64" means "int *", but a[1] is "int".

With this,
  ====
  p:probe_hello/foo /tmp/hello:0x4d3 a=+4(-8(%bp)):s32
  ====
So, "int" correctly converted to "s32"

Fixes: b2a3c12b7442 ("perf probe: Support tracing an entry of array")
Signed-off-by: Jan Stancek <jstancek@redhat.com>