[tools] perf annotate: Stop using a global config struct
Message-id: <20190211232742.16489-153-jolsa@redhat.com>
Patchwork-id: 241545
O-Subject: [PATCH RHEL7 BZ1672651 152/271] perf annotate: Stop using a global config struct
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 16932d77050fb3d76bc265c21c53eeec14639d5e
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Mar 15 12:33:56 2018 -0300

description
===========
For the TUI, that is interactive, its interesting to have a
configuration that one can go on changing and then when moving from one
symbol annotation to another symbol, the options set while browsing the
first symbol to be kept.

But since we're trying to make this code reusable by a --stdio
formatter, we better have a pointer in struct annotation and in the TUI
case set it to the global, but use something else for other cases, such
as --stdio2.

Signed-off-by: Jan Stancek <jstancek@redhat.com>