[tools] perf annotate: Defer searching for comma in raw line till it is needed
Message-id: <20190211232742.16489-187-jolsa@redhat.com>
Patchwork-id: 241571
O-Subject: [PATCH RHEL7 BZ1672651 186/271] perf annotate: Defer searching for comma in raw line till it is needed
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit c448234cfe46ec5abc0014dca8b3b49989bffe9e
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Mar 23 10:57:08 2018 -0300

description
===========
That strchr() in jump__scnprintf() needs to be nuked somehow, as it,
IIRC is already done in jump__parse() and if needed at scnprintf() time,
should be stashed in the struct filled in parse() time.

For now jus defer it to just before where it is used.

Signed-off-by: Jan Stancek <jstancek@redhat.com>