[tools] perf stat: Keep the / modifier separator in fallback
Message-id: <20190211232742.16489-251-jolsa@redhat.com>
Patchwork-id: 241644
O-Subject: [PATCH RHEL7 BZ1672651 250/271] perf stat: Keep the / modifier separator in fallback
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 129193bb0c43d42f1c397c175346e3e0dba5a578
Author: Jiri Olsa <jolsa@kernel.org>
Date: Mon Apr 23 11:08:17 2018 +0200

description
===========
The 'perf stat' fallback for EACCES error sets the exclude_kernel
perf_event_attr and tries perf_event_open() again with it. In addition,
it also changes the name of the event to reflect that change by adding
the 'u' modifier.

But it does not take into account the '/' separator, so the event name
can end up mangled, like: (note the '/:' characters)

  $ perf stat -e cpu/cpu-cycles/ kill
  ...
             386,832      cpu/cpu-cycles/:u

Adding the code to check on the '/' separator and set the following
correct event name:

  $ perf stat -e cpu/cpu-cycles/ kill
  ...
             388,548      cpu/cpu-cycles/u

Signed-off-by: Jan Stancek <jstancek@redhat.com>