[kernel] perf/core: Fix group scheduling with mixed hw and sw events
Message-id: <20190218130507.16170-2-jolsa@redhat.com>
Patchwork-id: 242029
O-Subject: [PATCH RHEL7 BZ1672893 001/187] perf/core: Fix group scheduling with mixed hw and sw events
Bugzilla: 1672893
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1672893
https://bugzilla.redhat.com/show_bug.cgi?id=1672893

upstream
========
commit a1150c202207cc8501bebc45b63c264f91959260
Author: Song Liu <songliubraving@fb.com>
Date: Thu May 3 12:47:16 2018 -0700

description
===========
When hw and sw events are mixed in the same group, they are all attached
to the hw perf_event_context. This sometimes requires moving group of
perf_event to a different context.

We found a bug in how the kernel handles this, for example if we do:

   perf stat -e '{faults,ref-cycles,faults}'  -I 1000

     1.005591180              1,297      faults
     1.005591180        457,476,576      ref-cycles
     1.005591180    <not supported>      faults

First, sw event "faults" is attached to the sw context, and becomes the
group leader. Then, hw event "ref-cycles" is attached, so both events
are moved to the hw context. Last, another sw "faults" tries to attach,
but it fails because of mismatch between the new target ctx (from sw
pmu) and the group_leader's ctx (hw context, same as ref-cycles).

The broken condition is:
   group_leader is sw event;
   group_leader is on hw context;
   add a sw event to the group.

Fix this scenario by checking group_leader's context (instead of just
event type). If group_leader is on hw context, use the ->pmu of this
context to look up context for the new event.

Fixes: b04243ef7006 ("perf: Complete software pmu grouping")
Signed-off-by: Jan Stancek <jstancek@redhat.com>