[tools] perf tools: Use the "_stest" symbol to identify the kernel map when loading kcore
Message-id: <20190218130507.16170-62-jolsa@redhat.com>
Patchwork-id: 242084
O-Subject: [PATCH RHEL7 BZ1672893 061/187] perf tools: Use the "_stest" symbol to identify the kernel map when loading kcore
Bugzilla: 1672893
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1672893
https://bugzilla.redhat.com/show_bug.cgi?id=1672893

upstream
========
commit 5654997838c2ac9b1950d633fc97f354cc4180e7
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed May 9 14:43:34 2018 +0300

description
===========
The first symbol is not necessarily in the kernel text.  Instead of
using the first symbol, use the _stest symbol to identify the kernel map
when loading kcore.

This allows for the introduction of symbols to identify the x86_64 PTI
entry trampolines.

Signed-off-by: Jan Stancek <jstancek@redhat.com>