[netdrv] qed: Lower the severity of a dcbx log message
Message-id: <1550520229-15616-62-git-send-email-mchopra@redhat.com>
Patchwork-id: 242346
O-Subject: [PATCH v2 RHEL 7.7 061/123] qed: Lower the severity of a dcbx log message.
Bugzilla: 1643532
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Kamal Heib <kheib@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

From: Sudarsana Reddy Kalluru <sudarsana.kalluru@cavium.com>

Bugzilla 1643532

commit c8b795669a5eb94fc97d2e1c967f0bad85d80046
https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git

Driver displays an error message for each unrecognized dcbx TLV that's
received from the peer or configured on the device. It is observed that
syslog will be flooded with such messages in certain scenarios e.g.,
frequent link-flaps/lldp-transactions. Changing the severity of this
message to verbose level as it's not an error scenario/message.

Signed-off-by: Sudarsana Reddy Kalluru <Sudarsana.Kalluru@cavium.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jan Stancek <jstancek@redhat.com>