[netdrv] qed: Fix blocking/unlimited SPQ entries leak
Message-id: <1550520229-15616-88-git-send-email-mchopra@redhat.com>
Patchwork-id: 242391
O-Subject: [PATCH v2 RHEL 7.7 087/123] qed: Fix blocking/unlimited SPQ entries leak
Bugzilla: 1643532
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Kamal Heib <kheib@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

From: Denis Bolotin <denis.bolotin@cavium.com>

Bugzilla 1643532

commit 2632f22ebd08da249c2017962a199a0cfb2324bf
https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git

When there are no SPQ entries left in the free_pool, new entries are
allocated and are added to the unlimited list. When an entry in the pool
is available, the content is copied from the original entry, and the new
entry is sent to the device. qed_spq_post() is not aware of that, so the
additional entry is stored in the original entry as p_post_ent, which can
later be returned to the pool.

Signed-off-by: Denis Bolotin <denis.bolotin@cavium.com>
Signed-off-by: Michal Kalderon <michal.kalderon@cavium.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jan Stancek <jstancek@redhat.com>