[netdrv] qed: Fix rdma_info structure allocation
Message-id: <1550520229-15616-93-git-send-email-mchopra@redhat.com>
Patchwork-id: 242412
O-Subject: [PATCH v2 RHEL 7.7 092/123] qed: Fix rdma_info structure allocation
Bugzilla: 1643532
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Kamal Heib <kheib@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

From: Michal Kalderon <michal.kalderon@cavium.com>

Bugzilla 1643532

commit 291d57f67d2449737d1e370ab5b9a583818eaa0c
https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git

Certain flows need to access the rdma-info structure, for example dcbx
update flows. In some cases there can be a race between the allocation or
deallocation of the structure which was done in roce start / roce stop and
an asynchrounous dcbx event that tries to access the structure.
For this reason, we move the allocation of the rdma_info structure to be
similar to the iscsi/fcoe info structures which are allocated during device
setup.
We add a new field of "active" to the struct to define whether roce has
already been started or not, and this is checked instead of whether the
pointer to the info structure.

Fixes: 51ff17251c9c ("qed: Add support for RoCE hw init")
Signed-off-by: Michal Kalderon <michal.kalderon@cavium.com>
Signed-off-by: Denis Bolotin <denis.bolotin@cavium.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jan Stancek <jstancek@redhat.com>