[tools] perf thread: Make thread__find_map() return the map
Message-id: <20190219091348.GA6121@krava>
Patchwork-id: 242420
O-Subject: [PATCHv2 RHEL7 BZ1672893 022/187] perf thread: Make thread__find_map() return the map
Bugzilla: 1672893
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672893
https://bugzilla.redhat.com/show_bug.cgi?id=1672893

upstream
========
commit 71a84b5aedf5023f4009c3bbf28ecba256201f87
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Apr 24 11:58:56 2018 -0300

description
===========
It was returning the searched map just on the addr_location passed, with
the function itself returning void.

Make it return the map so that we can make the code more compact.

RHEL7
=====
Ommited cs-etm.c hunk, because it's not backported in RHEL7.

Signed-off-by: Jan Stancek <jstancek@redhat.com>