[x86] kvm: vmx: Flush TLB when the APIC-access address changes
Message-id: <20190220180220.31125-23-vkuznets@redhat.com>
Patchwork-id: 242538
O-Subject: [RHEL7.7 KERNEL PATCH v2 022/121] kvm: vmx: Flush TLB when the APIC-access address changes
Bugzilla: 1565739 1497611
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>

From: Jim Mattson <jmattson@google.com>

Quoting from the Intel SDM, volume 3, section 28.3.3.4: Guidelines for
Use of the INVEPT Instruction:

If EPT was in use on a logical processor at one time with EPTP X, it
is recommended that software use the INVEPT instruction with the
"single-context" INVEPT type and with EPTP X in the INVEPT descriptor
before a VM entry on the same logical processor that enables EPT with
EPTP X and either (a) the "virtualize APIC accesses" VM-execution
control was changed from 0 to 1; or (b) the value of the APIC-access
address was changed.

In the nested case, the burden falls on L1, unless L0 enables EPT in
vmcs02 when L1 doesn't enable EPT in vmcs12.

Signed-off-by: Jim Mattson <jmattson@google.com>
Signed-off-by: Radim Krm <rkrcmar@redhat.com>
(cherry picked from commit fb6c8198431311027c3434d4e94ab8bc040f7aea)
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>

Conflicts:
	arch/x86/kvm/vmx.c (out of order ae1e2d1082 backport)