[netdrv] ice: fix changing of ring descriptor size (ethtool -G)
Message-id: <6334af25b2b7132335a3916928b9ec75426cbbcf.1550687071.git.jtoppins@redhat.com>
Patchwork-id: 242701
O-Subject: [PATCH RHEL7.7 061/149] ice: fix changing of ring descriptor size (ethtool -G)
Bugzilla: 1454916
RH-Acked-by: Honggang Li <honli@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

rx_mini_pending was set to an incorrect value. This was causing EINVAL to
always be returned to 'ethtool -G'. The driver does not support mini or
jumbo rings so the respective settings should be zero.

Also, change the valid range of the number of descriptors in the rings to
make the code simpler and easier for users to understand (this removes the
valid settings of 8 and 16). Add a system log message indicating when the
number is rounded-up from what the user specifies with the 'ethtool -G'
command (i.e. when it is not a multiple of 32), and update the log message
when a user-provided value is out of range to also indicate the stride.

Signed-off-by: Bruce Allan <bruce.w.allan@intel.com>
Signed-off-by: Anirudh Venkataramanan <anirudh.venkataramanan@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 3b6bf296c44f2b1233b7622c847a6ac5b1aeba0f)
Bugzilla: 1454916
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=20281962
Tested: boot & ping test done by Intel, no hardware in house
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>