[netdrv] ice: Change pf state behavior to protect reset path
Message-id: <9f4acd93784e784800e3e0b9b8d710d0dcebb879.1550687071.git.jtoppins@redhat.com>
Patchwork-id: 242641
O-Subject: [PATCH RHEL7.7 069/149] ice: Change pf state behavior to protect reset path
Bugzilla: 1454916
RH-Acked-by: Honggang Li <honli@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Currently, there is no bit, or set of bits, that protect the entirety
of the reset path.

If the reset is originated by the driver, then the relevant
one of the following bits will be set when the reset is scheduled:
__ICE_PFR_REQ
__ICE_CORER_REQ
__ICE_GLOBR_REQ
This bit will not be cleared until after the rebuild has completed.

If the reset is originated by the FW, then the first the driver knows of
it will be the reception of the OICR interrupt.  The __ICE_RESET_OICR_RECV
bit will be set in the interrupt handler.  This will also be the indicator
in a SW originated reset that we have completed the pre-OICR tasks and
have informed the FW that a reset was requested.

To utilize these bits, change the function:
ice_is_reset_recovery_pending()
	to be:
ice_is_reset_in_progress()

The new function will check all of the above bits in the pf->state and
will return a true if one or more of these bits are set.

Signed-off-by: Dave Ertman <david.m.ertman@intel.com>
Signed-off-by: Anirudh Venkataramanan <anirudh.venkataramanan@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 5df7e45d54fc99dd7c73e3a1f163cbfafc8b51f5)
Bugzilla: 1454916
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=20281962
Tested: boot & ping test done by Intel, no hardware in house
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>