[netdrv] qede: Correct XDP forward unmapping
Message-id: <1550868472-3818-5-git-send-email-mchopra@redhat.com>
Patchwork-id: 242963
O-Subject: [PATCH RHEL 7.7 04/14] qede: Correct XDP forward unmapping
Bugzilla: 1679753
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Bugzilla 1679753

commit 89e1afc44765d8b9f509d15df096494f14463e17
https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git

Driver is currently using dma_unmap_single() with the address it
passed to device for the purpose of forwarding, but the XDP
transmission buffer was originally a page allocated for the rx-queue.
The mapped address is likely to differ from the original mapped
address due to the placement offset.

This difference is going to get even bigger once we support headroom.

Cache the original mapped address of the page, and use it for unmapping
of the buffer when completion arrives for the XDP forwarded packet.

Backport note:
--------------
There are slight differences from upstream patch beacuse it is
being backported after various qede changes.

Signed-off-by: Yuval Mintz <Yuval.Mintz@cavium.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Manish Chopra <mchopra@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>