[netdrv] qede: Support XDP adjustment of headers
Message-id: <1550868472-3818-8-git-send-email-mchopra@redhat.com>
Patchwork-id: 242966
O-Subject: [PATCH RHEL 7.7 07/14] qede: Support XDP adjustment of headers
Bugzilla: 1679753
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Bugzilla 1679753

commit 059eeb07e175086db1f84c1d8d29bb9aa8057797
https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git

In case an XDP program is attached, reserve XDP_PACKET_HEADROOM
bytes at the beginning of the packet for the program to play
with.

Modify the XDP logic in the driver to fill-in the missing bits
and re-calculate offsets and length after the program has finished
running to properly reflect the current status of the packet.

We can then go and remove the limitation of not supporting XDP programs
where xdp_adjust_head is set.

Backport note:
--------------
Slight differences from upstream as backport after lot of qede patches.

Signed-off-by: Yuval Mintz <Yuval.Mintz@cavium.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Manish Chopra <mchopra@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>