[netdrv] xdp: make the stack take care of the tear down
Message-id: <1550868472-3818-14-git-send-email-mchopra@redhat.com>
Patchwork-id: 242971
O-Subject: [PATCH RHEL 7.7 13/14] net: xdp: make the stack take care of the tear down
Bugzilla: 1679753
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Bugzilla 1679753

commit bd0b2e7fe611953470ec7c533b455fb2abd382cd
https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git

Since day one of XDP drivers had to remember to free the program
on the remove path.  This leads to code duplication and is error
prone.  Make the stack query the installed programs on unregister
and if something is installed, remove the program.  Freeing of
program attached to XDP generic is moved from free_netdev() as well.

Because the remove will now be called before notifiers are
invoked, BPF offload state of the program will not get destroyed
before uninstall.

Backport note:
--------------
Taken qede slice which was dropped from earlier same patch backport.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Simon Horman <simon.horman@netronome.com>
Reviewed-by: Quentin Monnet <quentin.monnet@netronome.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Manish Chopra <mchopra@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>