[tools] tools include uapi: Grab a copy of linux/in.h
Message-id: <20190226093100.17420-35-jolsa@redhat.com>
Patchwork-id: 243161
O-Subject: [PATCH RHEL7 BZ1680072 034/111] tools include uapi: Grab a copy of linux/in.h
Bugzilla: 1680072
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1680072
https://bugzilla.redhat.com/show_bug.cgi?id=1680072

upstream
========
commit a4b2061242ecc024b1ee975e066b6e38db1c6b9e
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Jul 26 09:26:13 2018 -0300

description
===========
We'll use it to create tables for the 'protocol' argument to the
socket syscall when the 'family' arg is one of AF_INET or AF_INET6.

Add it to check_headers.sh so that when a new protocol gets added we get
a notification during the build process.

RHEL7
=====
Taking rhel7 copy of linux/in.h.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>