[tools] perf trace: Associate vfs_getname()'ed pathname with fd returned from 'openat'
Message-id: <20190226093100.17420-47-jolsa@redhat.com>
Patchwork-id: 243174
O-Subject: [PATCH RHEL7 BZ1680072 046/111] perf trace: Associate vfs_getname()'ed pathname with fd returned from 'openat'
Bugzilla: 1680072
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1680072
https://bugzilla.redhat.com/show_bug.cgi?id=1680072

upstream
========
commit 6a648b534d3d4f6532241fd9b97c0dab7b0c3f89
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Aug 2 10:30:08 2018 -0300

description
===========
When the vfs_getname() wannabe tracepoint is in place:

  # perf probe -l
    probe:vfs_getname    (on getname_flags:73@acme/git/linux/fs/namei.c with pathname)
  #

'perf trace' will use it to get the pathname when it is copied from
userspace to the kernel, right after syscalls:sys_enter_open, copied
in the 'probe:vfs_getname', stash it somewhere and then, at
syscalls:sys_exit_open time, if the 'open' return is not -1, i.e. a
successfull open syscall, associate that pathname to this return, i.e.
the fd.

We were not doing this for the 'openat' syscall, which would cause 'perf
trace' to fallback to using /proc to get the fd, change it so that we
use what we got from probe:vfs_getname, reducing the 'openat'
beautification process cost, ditching the syscalls performed to read
procfs state and avoiding some possible races in the process.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>