[tools] perf parser: Improve error message for PMU address filters
Message-id: <20190226093100.17420-79-jolsa@redhat.com>
Patchwork-id: 243209
O-Subject: [PATCH RHEL7 BZ1680072 078/111] perf parser: Improve error message for PMU address filters
Bugzilla: 1680072
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1680072
https://bugzilla.redhat.com/show_bug.cgi?id=1680072

upstream
========
commit 49836f7811f383d8613661fff110ce74f4710d52
Author: Jack Henschel <jackdev@mailbox.org>
Date: Wed Jul 4 14:13:45 2018 +0200

description
===========
This is the second version of a patch that improves the error message of
the perf events parser when the PMU hardware does not support address
filters.

Previously, the perf returned the following error:

  $ perf record -e intel_pt// --filter 'filter sys_write'
  --filter option should follow a -e tracepoint or HW tracer option

This implies there is some syntax error present in the command line,
which is not true. Rather, notify the user that the CPU does not have
support for this feature.

For example, Intel chips based on the Broadwell micro-archticture have
the Intel PT PMU, but do not support address filtering.

Now, perf prints the following error message:

  $ perf record -e intel_pt// --filter 'filter sys_write'
  This CPU does not support address filtering

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>