[tools] perf tools: Move the temp file processing into decompress_kmodule
Message-id: <20190226093100.17420-88-jolsa@redhat.com>
Patchwork-id: 243222
O-Subject: [PATCH RHEL7 BZ1680072 087/111] perf tools: Move the temp file processing into decompress_kmodule
Bugzilla: 1680072
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1680072
https://bugzilla.redhat.com/show_bug.cgi?id=1680072

upstream
========
commit c9a8a6131fb68f6eabec34fd855e69b4855155e4
Author: Jiri Olsa <jolsa@kernel.org>
Date: Fri Aug 17 11:48:09 2018 +0200

description
===========
We will add a compression check in the following patch and it makes it
easier if the file processing is done in a single place. It also makes
the current code simpler.

The decompress_kmodule function now returns the fd of the uncompressed
file and the file name in the pathname arg, if it's provided.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>