[tools] perf tools: Add is_compressed callback to compressions array
Message-id: <20190226093100.17420-89-jolsa@redhat.com>
Patchwork-id: 243233
O-Subject: [PATCH RHEL7 BZ1680072 088/111] perf tools: Add is_compressed callback to compressions array
Bugzilla: 1680072
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1680072
https://bugzilla.redhat.com/show_bug.cgi?id=1680072

upstream
========
commit 8b42b7e5e8b5692bc5e57342bf00c40d64978407
Author: Jiri Olsa <jolsa@kernel.org>
Date: Fri Aug 17 11:48:10 2018 +0200

description
===========
Add is_compressed callback to the compressions array, that returns 0 if
the file is compressed or != 0 if not.

The new callback is used to recognize the situation when we have a
'compressed' object, like:

  /lib/modules/.../drivers/net/ethernet/intel/igb/igb.ko.xz

but we need to read its debug data from debuginfo files, which might not
be compressed, like:

  /root/.debug/.build-id/d6/...c4b301f/debug

So even for a 'compressed' object we read debug data from a plain
uncompressed object. To keep this transparent, we detect this in
decompress_kmodule() and return the file descriptor to the uncompressed
file.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>