[tools] perf util: Fix bad memory access in trace info
Message-id: <20190226093100.17420-95-jolsa@redhat.com>
Patchwork-id: 243219
O-Subject: [PATCH RHEL7 BZ1680072 094/111] perf util: Fix bad memory access in trace info.
Bugzilla: 1680072
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1680072
https://bugzilla.redhat.com/show_bug.cgi?id=1680072

upstream
========
commit a72f64261359b7451f8478f2a2bf357b4e6c757f
Author: Chris Phlipot <cphlipot0@gmail.com>
Date: Tue Aug 28 23:19:54 2018 -0700

description
===========
In the write to the output_fd in the error condition of
record_saved_cmdline(), we are writing 8 bytes from a memory location on
the stack that contains a primitive that is only 4 bytes in size.
Change the primitive to 8 bytes in size to match the size of the write
in order to avoid reading unknown memory from the stack.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>