[tools] perf tools: Fix use of alternatives to find JDIR
Message-id: <20190226093100.17420-108-jolsa@redhat.com>
Patchwork-id: 243218
O-Subject: [PATCH RHEL7 BZ1680072 107/111] perf tools: Fix use of alternatives to find JDIR
Bugzilla: 1680072
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1680072
https://bugzilla.redhat.com/show_bug.cgi?id=1680072

upstream
========
commit 36b8d4628d3cc8f5a748e508cce8673bc00fc63c
Author: Jarod Wilson <jarod@redhat.com>
Date: Thu Sep 6 18:18:12 2018 -0400

description
===========
When a build is run from something like a cron job, the user's $PATH is
rather minimal, of note, not including /usr/sbin in my own case. Because
of that, an automated rpm package build ultimately fails to find
libperf-jvmti.so, because somewhere within the build, this happens...

  /bin/sh: alternatives: command not found
  /bin/sh: alternatives: command not found
  Makefile.config:849: No openjdk development package found, please install
  JDK package, e.g. openjdk-8-jdk, java-1.8.0-openjdk-devel

...and while the build continues, libperf-jvmti.so isn't built, and
things fall down when rpm tries to find all the %files specified. Exact
same system builds everything just fine when the job is launched from a
login shell instead of a cron job, since alternatives is in $PATH, so
openjdk is actually found.

The test required to get into this section of code actually specifies
the full path, as does a block just above it, so let's do that here too.

Fixes: d4dfdf00d43e ("perf jvmti: Plug compilation into perf build")
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>