[lib] rhashtable: Cap total number of entries to 2^31
Message-id: <1551175657-23814-8-git-send-email-crecklin@redhat.com>
Patchwork-id: 243253
O-Subject: [RHEL7.7 BZ1666812 07/12] rhashtable: Cap total number of entries to 2^31
Bugzilla: 1666812
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Joe Lawrence <joe.lawrence@redhat.com>

From: Herbert Xu <herbert@gondor.apana.org.au>

When max_size is not set or if it set to a sufficiently large
value, the nelems counter can overflow.  This would cause havoc
with the automatic shrinking as it would then attempt to fit a
huge number of entries into a tiny hash table.

This patch fixes this by adding max_elems to struct rhashtable
to cap the number of elements.  This is set to 2^31 as nelems is
not a precise count.  This is sufficiently smaller than UINT_MAX
that it should be safe.

When max_size is set max_elems will be lowered to at most twice
max_size as is the status quo.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6d684e54690caef45cf14051ddeb7c71beeb681b)
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>