[fs] ext4: check for NUL characters in extended attribute's name
Message-id: <20190301141859.5418-25-lczerner@redhat.com>
Patchwork-id: 243799
O-Subject: [RHEL7.7 fs PATCH v2 24/51] ext4: check for NUL characters in extended attribute's name
Bugzilla: 1671293
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

From: Theodore Ts'o <tytso@mit.edu>

BZ 1671293
https://bugzilla.redhat.com/show_bug.cgi?id=1671293

BREW 20414283
https://brewweb.devel.redhat.com/taskinfo?taskID=20414283

commit 7d95178c77014dbd8dce36ee40bbbc5e6c121ff5

Extended attribute names are defined to be NUL-terminated, so the name
must not contain a NUL character.  This is important because there are
places when remove extended attribute, the code uses strlen to
determine the length of the entry.  That should probably be fixed at
some point, but code is currently really messy, so the simplest fix
for now is to simply validate that the extended attributes are sane.

https://bugzilla.kernel.org/show_bug.cgi?id=200401

Reported-by: Wen Xu <wen.xu@gatech.edu>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: stable@vger.kernel.org
---

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>