[fs] ext4: avoid kernel warning when writing the superblock to a dead device
Message-id: <20190301141859.5418-50-lczerner@redhat.com>
Patchwork-id: 243829
O-Subject: [RHEL7.7 fs PATCH v2 49/51] ext4: avoid kernel warning when writing the superblock to a dead device
Bugzilla: 1671293
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

From: Theodore Ts'o <tytso@mit.edu>

BZ 1671293
https://bugzilla.redhat.com/show_bug.cgi?id=1671293

BREW 20414283
https://brewweb.devel.redhat.com/taskinfo?taskID=20414283

commit e86807862e6880809f191c4cea7f88a489f0ed34

The xfstests generic/475 test switches the underlying device with
dm-error while running a stress test.  This results in a large number
of file system errors, and since we can't lock the buffer head when
marking the superblock dirty in the ext4_grp_locked_error() case, it's
possible the superblock to be !buffer_uptodate() without
buffer_write_io_error() being true.

We need to set buffer_uptodate() before we call mark_buffer_dirty() or
this will trigger a WARN_ON.  It's safe to do this since the
superblock must have been properly read into memory or the mount would
have been successful.  So if buffer_uptodate() is not set, we can
safely assume that this happened due to a failed attempt to write the
superblock.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: stable@vger.kernel.org
---

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>