[sound] alsa: virmidi: Fix too long output trigger loop
Message-id: <15517824857704096-27-patchset-jkysela@redhat.com>
Patchwork-id: 244102
O-Subject: [RHEL 7.7 PATCH 026/467] ALSA: virmidi: Fix too long output trigger loop
Bugzilla: 1657855
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The virmidi output trigger tries to parse the all available bytes and
process sequencer events as much as possible.  In a normal situation,
this is supposed to be relatively short, but a program may give a huge
buffer and it'll take a long time in a single spin lock, which may
eventually lead to a soft lockup.

This patch simply adds a workaround, a cond_resched() call in the loop
if applicable.  A better solution would be to move the event processor
into a work, but let's put a duct-tape quickly at first.

Reported-and-tested-by: Dae R. Jeong <threeearcat@gmail.com>
Reported-by: syzbot+619d9f40141d826b097e@syzkaller.appspotmail.com
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 50e9ffb1996a5d11ff5040a266585bad4ceeca0a)
Bugzilla: 1657855
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>