[sound] alsa: usb-audio: UAC3. Add support for mixer unit
Message-id: <15517824857704096-82-patchset-jkysela@redhat.com>
Patchwork-id: 244154
O-Subject: [RHEL 7.7 PATCH 081/467] ALSA: usb-audio: UAC3. Add support for mixer unit.
Bugzilla: 1657855
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This adds support for the MIXER UNIT in UAC3. All the information
is obtained from the (HIGH CAPABILITY) Cluster's header. We don't
read the rest of the logical cluster to obtain the channel config
as that wont make any difference in the current mixer behaviour.

The name of the mixer unit is not yet requested as there is not
support for the UAC3 Class Specific String requests.

Tested in an UAC3 device working as a HEADSET with a basic mixer
unit (same as the one in the BADD spec) with no controls.

Signed-off-by: Jorge Sanjuan <jorge.sanjuan@codethink.co.uk>
Reviewed-by: Ruslan Bilovol <ruslan.bilovol@gmail.com>
Tested-by: Ruslan Bilovol <ruslan.bilovol@gmail.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 6cfd839ae78ec3fac5ddbf7148155898727e90c3)
Bugzilla: 1657855
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>