[sound] alsa: hda/ca0132: Don't test for QUIRK_NONE
Message-id: <15517824857704096-188-patchset-jkysela@redhat.com>
Patchwork-id: 244270
O-Subject: [RHEL 7.7 PATCH 187/467] ALSA: hda/ca0132: Don't test for QUIRK_NONE
Bugzilla: 1657855
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

QUIRK_NONE is, quite explicitly, the default case.  The entire
point of a quirks system is to allow "programming by difference"
from a given base case, which requires that merely defining a new
quirk for some piece of hardware should not change the behavior of
the driver for that hardware.  In turn, this means that testing
for QUIRK_NONE explicitly is a violation of that implicit contract.

Change a test for QUIRK_NONE and QUIRK_ALIENWARE to default, and
add a test for QUIRK_SBZ to disable the default behavior in that
instance.

Signed-off-by: Alastair Bridgewater <alastair.bridgewater@gmail.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 126b75e03894f3053b723130cb9b824e6f15d8e6)
Bugzilla: 1657855
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>