[hwtracing] stm class: Use driver's packet callback return value
Message-id: <20190311092113.21814-15-jolsa@redhat.com>
Patchwork-id: 245632
O-Subject: [PATCH RHEL7 BZ1514700 014/125] stm class: Use driver's packet callback return value
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit f8560a9bc76b2cd5c06fa412cb7b5481d70fcf34
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Mon Feb 15 19:12:01 2016 +0200

description
===========
STM drivers provide a callback to generate/send individual STP packets;
it also tells the stm core how many bytes of payload it has consumed.
However, we would also need to use the negative space of this return
value to communicate errors that occur during the packet generation,
in which case the stm core will have to take appropriate action.

For now, we need to account for the possibility that the stm driver may
not support certain combinations of packet type/flags, in which case
it is expected to signal an error.

Signed-off-by: Jan Stancek <jstancek@redhat.com>