[hwtracing] stm class: Fix unlocking braino in the error path
Message-id: <20190311092113.21814-19-jolsa@redhat.com>
Patchwork-id: 245619
O-Subject: [PATCH RHEL7 BZ1514700 018/125] stm class: Fix unlocking braino in the error path
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit 1810f2c44817c74ca3d05d1e3981e3a2e2ceb6f5
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Mon Feb 15 19:12:05 2016 +0200

description
===========
If an illegal attempt is made to unlink stm source device from an
stm device, the stm device's link spinlock mistakenly remains locked.
While this really shouldn't happen (there's a warning in place), the
locking should remain in order so that we can still recover from this
situation if it indeed does happen.

This patch unifies the unlocking in the exit path of
__stm_source_link_drop() to fix this.

Signed-off-by: Jan Stancek <jstancek@redhat.com>