[hwtracing] stm class: Fix a race in unlinking
Message-id: <20190311092113.21814-22-jolsa@redhat.com>
Patchwork-id: 245629
O-Subject: [PATCH RHEL7 BZ1514700 021/125] stm class: Fix a race in unlinking
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit b4ca34aaf78ed0cdfc15956d377064104257a437
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Mon Feb 15 19:12:08 2016 +0200

description
===========
There is a window in stm_source_link_drop(), during which the source's
link may change before locks are acquired. When this happens, it throws
a warning, since this is not an expected scenario.

This patch handles the race in such a way that if the link appears to
have changed by the time we took the locks, it will release them and
repeat the whole unlinking procedure from the beginning, unless the
other contender beat us to it.

Signed-off-by: Jan Stancek <jstancek@redhat.com>