[hwtracing] intel_th: Support Host Debugger mode of operation
Message-id: <20190311092113.21814-73-jolsa@redhat.com>
Patchwork-id: 245692
O-Subject: [PATCH RHEL7 BZ1514700 072/125] intel_th: Support Host Debugger mode of operation
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit c49a75910c5ea9bbeb60a86350f232f6fcb13cc2
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Mon Sep 19 17:07:47 2016 +0300

description
===========
This patch adds a 'host_mode' module option to enable host-driven
operational mode in the driver. In this mode, the driver does not
perform trace configuration or enable trace capture, but still
provides all the means necessary for software trace sources to
write their data to the Trace Hub. This means that the debug host
takes care of all the configuration and enabling and we do not
interfere.

Signed-off-by: Jan Stancek <jstancek@redhat.com>