[hwtracing] intel_th: Don't leak module refcount on failure to activate
Message-id: <20190311092113.21814-76-jolsa@redhat.com>
Patchwork-id: 245675
O-Subject: [PATCH RHEL7 BZ1514700 075/125] intel_th: Don't leak module refcount on failure to activate
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit e609ccef5222c73b46b322be7d3796d60bff353d
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri Feb 24 16:04:15 2017 +0200

description
===========
Output 'activation' may fail for the reasons of the output driver,
for example, if msc's buffer is not allocated. We forget, however,
to drop the module reference in this case. So each attempt at
activation in this case leaks a reference, preventing the module
from ever unloading.

This patch adds the missing module_put() in the activation error
path.

Signed-off-by: Jan Stancek <jstancek@redhat.com>