[hwtracing] intel_th: Perform time resync on capture start
Message-id: <20190311092113.21814-92-jolsa@redhat.com>
Patchwork-id: 245689
O-Subject: [PATCH RHEL7 BZ1514700 091/125] intel_th: Perform time resync on capture start
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit a0e7df335afd2a8a8a688251ffee375b58b6517c
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri Feb 24 16:09:40 2017 +0200

description
===========
On some devices (TH 2.x devices at the moment), the internal time counter
is initially not synchronized to the global crystal clock, so the time
stamps it produces will not be useful. In this case, the driver needs
to force the time counter resync.

This applies the workaround to relevant devices.

Signed-off-by: Jan Stancek <jstancek@redhat.com>