[hwtracing] intel_th: Don't touch switch routing in host mode
Message-id: <20190311092113.21814-97-jolsa@redhat.com>
Patchwork-id: 245697
O-Subject: [PATCH RHEL7 BZ1514700 096/125] intel_th: Don't touch switch routing in host mode
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit c2d2c7de972d7a08533bde622c36ff6feb112907
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Thu Mar 1 10:17:25 2018 +0200

description
===========
When the Trace Hub is operating in Host Debugger mode, it is up to the
debugger to configure master routing even for the software sources. Do
not do this in the driver in this case.

Signed-off-by: Jan Stancek <jstancek@redhat.com>