[hwtracing] intel_th: Fix resource handling for ACPI glue layer
Message-id: <20190311092113.21814-103-jolsa@redhat.com>
Patchwork-id: 245717
O-Subject: [PATCH RHEL7 BZ1514700 102/125] intel_th: Fix resource handling for ACPI glue layer
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit ebe4582281d6e90972f057318a6edea14810ea48
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Tue Sep 18 16:10:48 2018 +0300

description
===========
The core of the driver expects the resource array from the glue layer
to be indexed by even numbers, as is the case for 64-bit PCI resources.
This doesn't hold true for others, ACPI in this instance, which leads
to an out-of-bounds access and an ioremap() on whatever address that
access fetches.

This patch fixes the problem by reading resource array differently based
on whether the 64-bit flag is set, which would indicate PCI glue layer.

Fixes: ebc57e399b8e ("intel_th: Add ACPI glue layer")
Signed-off-by: Jan Stancek <jstancek@redhat.com>