[hwtracing] stm class: Fix a use-after-free
Message-id: <20190311092113.21814-107-jolsa@redhat.com>
Patchwork-id: 245706
O-Subject: [PATCH RHEL7 BZ1514700 106/125] stm class: Fix a use-after-free
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit fd085bb1766d6a598f53af2308374a546a49775a
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Tue Sep 19 18:47:40 2017 +0300

description
===========
For reasons unknown, the stm_source removal path uses device_destroy()
to kill the underlying device object. Because device_destroy() uses
devt to look for the device to destroy and the fact that stm_source
devices don't have one (or all have the same one), it just picks the
first device in the class, which may well be the wrong one.

That is, loading stm_console and stm_heartbeat and then removing both
will die in dereferencing a freed object.

Since this should have been device_unregister() in the first place,
use it instead of device_destroy().

Fixes: 7bd1d4093c2 ("stm class: Introduce an abstraction for System Trace Module devices")
Signed-off-by: Jan Stancek <jstancek@redhat.com>