[hwtracing] stm class: Rework policy node fallback
Message-id: <20190311092113.21814-111-jolsa@redhat.com>
Patchwork-id: 245708
O-Subject: [PATCH RHEL7 BZ1514700 110/125] stm class: Rework policy node fallback
Bugzilla: 1514700
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1514700
https://bugzilla.redhat.com/show_bug.cgi?id=1514700

upstream
========
commit cb6102bd99efe35f016dc6d7282e681e6dbde154
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Fri Oct 5 15:42:51 2018 +0300

description
===========
Currently, if no matching policy node can be found for a trace source,
we'll try to use "default" policy node, then, if that doesn't exist,
we'll pick the first node, in order of creation. If that also fails,
we'll allocate M/C range from the beginning of the device's M/C range.

This makes it difficult to know which node (if any) was used in any
particular case.

In order to make things more deterministic, the new order is as follows:
  * if they supply ID string, use that and nothing else,
  * if they are a task, use their task name (comm),
  * use "default", if it exists,
  * return failure, to let them know there is no suitable rule.

This should provide enough convenience with the "default" catch-all node,
while not leaving *everything* to chance. As a side effect, this relaxes
the requirement of using ioctl() for identification with the possibility of
using task names as policy nodes.

Signed-off-by: Jan Stancek <jstancek@redhat.com>